requirements.txt
alembic==1.4.2
astroid==2.4.1
autopep8==1.5.2
blinker==1.4
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
dnspython==1.16.0
email-validator==1.1.0
Flask==1.1.2
Flask-Admin==1.5.6
Flask-Dance==3.0.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.1
Flask-Uploads==0.2.1
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
Mako==1.1.2
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.1.0
Pillow==7.1.2
pycodestyle==2.6.0
pylint==2.5.2
python-dateutil==2.8.1
python-editor==1.0.4
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
SQLAlchemy==1.3.17
SQLAlchemy-Utils==0.36.5
stripe==2.48.0
toml==0.10.1
urllib3==1.25.9
URLObject==2.4.3
uWSGI==2.0.18
Werkzeug==0.16.0
wrapt==1.12.1
WTForms==2.3.1
